+++ knockturn-alley-driver/CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2023-10-27
+### Added
+- `Legilimens` spell now also gathers data from Neighbors Table (LQI), Routing Table and Bindings Table
+
 ## [1.4.0] - 2023-10-26
 ### Added
-- `Legilimens` spell now also gathers data from "Node Descriptor" and "Node Power Descriptor"
+- `Legilimens` spell now also gathers data from Node Descriptor and Node Power Descriptor
 - Translate attribute hex value to friendly representations for some known attributes (e.g. Power On Behavior, Temperature, Relative Humidity, etc.)
 
 ## [1.3.0] - 2023-10-04